Basic functionality works The app fetches json from the server, it parses it and creates a data model. From there on, it handles the data flow by the domain layer, which exposes the data model to the bloc. The BLoC handles the request of GreenHouseEvent.load from the UI and asks the domain for the data model. Then the UI displays the right values.
Note: there's the necessary to refactor a bit the UI. Moreover, I was thinking of
refactor the data model, maybe using generics and the repository layer, making it
a bit more scalable. Even though it's a really simple app, I want it to be as
scalable as possibile so that I could add features in the near future.